Move page code around in files, first step; tests pass
